[feat/prompt-plugin] feat: Add searchParams to ThemeComponent
This commit adds the  hook from Next.js to the  in the  file. It also includes the  as a query parameter when navigating to the  route. This change improves the functionality of the theme screen by allowing the  to be passed as a query parameter to the game screen.

Note: The commit message has been generated based on the provided code changes and recent repository commits.